Current version has (mainly): - periodic send of beacon with dinamic changing of BO and SO; - firs GTS management (allocation is decided only by the coord, no negotiation); - fixes on the radio driver; - creation of the abstract list (fixed size) implemented as array with   "iterators" some functions: just tested add, append, iterate (remove missing) - make the cqueue a "allocator" version, it works with pointer to internal   storage instead of doing memcpying.
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@98 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ contrib/network/openzb/inc/mac/ozb_mac_mlme.h
@@ -125,7 +125,7 @@ int8_t ozb_MLME_SET_confirm(enum ozb_mac_code_t status,
         /*TODO: enough?*/   uint16_t PIBAttributeIndex);
 
 #ifndef OZB_RFD_DISABLE_OPTIONAL
-int8_t ozb_MLME_START_request(uint8_t PANId, uint8_t LogicalChannel,
+int8_t ozb_MLME_START_request(uint16_t PANId, uint8_t LogicalChannel,
 			      uint8_t ChannelPage, uint32_t StartTime,
 			      uint8_t BeaconOrder, uint8_t SuperframeOrder,
 			      uint8_t PANCoordinator,
